Implement can delete only if genre has no books, refactoring, and fix some bugs

+++ code/src/Controller/GenreController.php
@@ -83,6 +83,10 @@ class GenreController extends AbstractController
      */
     public function delete(Request $request, Genre $genre): Response
     {
+        if ($genre->getLivres()->count() > 0)
+        {
+            return $this->redirectToRoute('genre_index', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$genre->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($genre);
